fabric: Define mechanism to support tagged atomic operations
Atomic operations are currently most closely related to RMA
operations.  Memory registered for RMA may also be the target of
atomic operations.  However, there's no reason why an atomic
operation couldn't target a tagged memory buffer as well.  This
can be done by treating the address as an offset into the tagged
buffer and the key as the tag.

Fixes #2557

Note that we don't define a mechanism for issuing atomics against
normal receive buffers because there have not been any application
requirements to support this.
