+++ include/rdma/fi_domain.h
@@ -108,9 +108,59 @@ struct fi_mr_attr {
 };
 
 
+#ifdef FABRIC_DIRECT
+#include <rdma/fi_direct_atomic_def.h>
+#endif /* FABRIC_DIRECT */
+
+#ifndef FABRIC_DIRECT_ATOMIC_DEF
+
+enum fi_datatype {
+	FI_INT8,
+	FI_UINT8,
+	FI_INT16,
+	FI_UINT16,
+	FI_INT32,
+	FI_UINT32,
+	FI_INT64,
+	FI_UINT64,
+	FI_FLOAT,
+	FI_DOUBLE,
+	FI_FLOAT_COMPLEX,
+	FI_DOUBLE_COMPLEX,
+	FI_LONG_DOUBLE,
+	FI_LONG_DOUBLE_COMPLEX,
+	FI_DATATYPE_LAST
+};
+
+enum fi_op {
+	FI_MIN,
+	FI_MAX,
+	FI_SUM,
+	FI_PROD,
+	FI_LOR,
+	FI_LAND,
+	FI_BOR,
+	FI_BAND,
+	FI_LXOR,
+	FI_BXOR,
+	FI_ATOMIC_READ,
+	FI_ATOMIC_WRITE,
+	FI_CSWAP,
+	FI_CSWAP_NE,
+	FI_CSWAP_LE,
+	FI_CSWAP_LT,
+	FI_CSWAP_GE,
+	FI_CSWAP_GT,
+	FI_MSWAP,
+	FI_ATOMIC_OP_LAST
+};
+
+#endif
+
+
+struct fi_atomic_attr;
 struct fi_cq_attr;
 struct fi_cntr_attr;
-
 
 struct fi_ops_domain {
 	size_t	size;
@@ -132,6 +182,9 @@ struct fi_ops_domain {
 	int	(*srx_ctx)(struct fid_domain *domain,
 			struct fi_rx_attr *attr, struct fid_ep **rx_ep,
 			void *context);
+	int	(*query_atomic)(struct fid_domain *domain,
+			enum fi_datatype datatype, enum fi_op op,
+			struct fi_atomic_attr *attr, uint64_t flags);
 };
